backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic>=2
python-multipart
python-jose[cryptography]
passlib>=1.7.4
bcrypt==3.2.0
python-dotenv
pandas
numpy
matplotlib
openpyxl
Jinja2
email-validator
cryptography
xlsxwriter
requests